// frontend/src/types/auth.ts
export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  type: string;
}

export interface User {
  email: string;
  firstName: string;
  lastName: string;
  role: string;
}